requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
alembic>=1.13,<2
pydantic>=2.7,<3
httpx>=0.27,<1
pytest>=7,<9
PyYAML>=6,<7
